Move version to app module
